log_group/requirements.txt
aws-cdk.core
aws_cdk.aws_logs